create slocum submodules
